+++ resources/views/backend/role/role/list.blade.php
@@ -5,10 +5,11 @@
             <h2 class="font-semibold text-xl text-gray-800  leading-tight">
                 {{ __('Roles') }}
             </h2>
-
-            <a href="{{route('roles.create')}}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">
-                Create
-            </a>
+            @can('create roles')
+                <a href="{{route('roles.create')}}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">
+                    Create
+                </a>
+            @endcan
         </div>
     </x-slot>
 
@@ -45,13 +46,17 @@
                                     </td>
                                     <td class="px-6 py-3 text-left">{{\Carbon\Carbon::parse($role->created_at)->format('d M, Y')}}</td>
                                     <td class="px-6 py-3 text-center">
-                                        <a href="{{route('roles.edit', $role->id)}}" class="bg-blue-700 text-sm rounded-md px-3 py-2 text-white hover:bg-blue-500">
-                                            Edit
-                                        </a>
+                                        @can('destroy roles')
+                                            <a href="{{route('roles.edit', $role->id)}}" class="bg-blue-700 text-sm rounded-md px-3 py-2 text-white hover:bg-blue-500">
+                                                Edit
+                                            </a>
+                                        @endcan
 
-                                        <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-700 text-sm rounded-md px-3 py-2 text-white hover:bg-red-500">
-                                            Delete
-                                        </a>
+                                        @can('destroy roles')
+                                            <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-700 text-sm rounded-md px-3 py-2 text-white hover:bg-red-500">
+                                                Delete
+                                            </a>
+                                        @endcan
                                     </td>
                                 </tr>
                             @endforeach
